requirements-docker.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==2.0.0
sqlalchemy==2.0.15
alembic==1.11.1
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
docker==6.1.2
pyyaml==6.0
redis==4.5.5
prometheus-client==0.17.0
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.1
sympy==1.12
networkx==2.8.0
pandas==1.5.0
matplotlib==3.5.0
pytest==7.3.1
pytest-asyncio==0.21.0
cryptography==41.0.0
jupyter==1.0.0
plotly==5.0.0
jinja2==3.0.0